Change names of modules + call projection code
