+++ src/task1/filter_module/images/inactive.jpg
