+++ src/task1/mark-damaged-module/result.json
@@ -1,1 +1,1 @@
-{"../common/test/2018-05-25_16-22-29-018.bmp": [[408, 304], [596, 302], [619, 158], [472, 188], [874, 256], [738, 392]], "../common/test/2018-05-25_16-22-29-685.bmp": []}
+[{"gps": [[43.75011648730608, -79.47143038799156], [43.75012407540277, -79.47142066404253]], "pixel": [[362, 164], [594, 194]], "file": "C:\\Users\\Jing\\Documents\\Work\\Other\\USC2019\\vision_201819\\src\\task1\\common\\test\\2018-05-25_16-22-29-018.bmp"}, {"gps": [[43.750202672464596, -79.47160374736728], [43.75021122490616, -79.47159929065009], [43.750203917141675, -79.47159729098958], [43.75020641644452, -79.47160606725191]], "pixel": [[343, 282], [531, 224], [428, 345], [383, 204]], "file": "C:\\Users\\Jing\\Documents\\Work\\Other\\USC2019\\vision_201819\\src\\task1\\common\\test\\2018-05-25_16-22-29-685.bmp"}]
